Port qoshdim va index.js papkasiga ozgartirishlar kiritdim

+++ Controller/bookControllers.js
@@ -1,5 +1,5 @@
 const mongoose = require("mongoose");
-
+const { getAuthorById } = require("./authorControllers");
 const bookSchema = {
   name: String,
   authors: [
@@ -18,19 +18,54 @@ async function getBookById(id) {
   return book;
 }
 
-async function updateBook(id, data) {
-  const { name, authors } = data;
+async function getBooks(req, res) {
+  const authors = await BookModel.find().populate("authors");
 
-  const updatedBook = await BookModel.updateOne({ id }, { name, authors });
-
-  return updatedBook;
+  res.status(201).send(authors);
 }
 
-async function deleteBook(id) {
+async function updateBook(req, res) {
+  const { id } = req.params;
+  const { name } = req.body;
+
+  const book = await getBookById(id);
+
+  if (!book) {
+    return res.status(404).send("Book not found!");
+  }
+  const updatedBook = await BookModel.updateOne({ id }, { name });
+
+  return res
+    .status(201)
+    .send({ msg: "Successfully updated", data: updatedBook });
+}
+
+async function deleteBook(req, res) {
+  const { id } = req.params;
+
+  const book = await getBookById(id);
+
+  if (!book) {
+    return res.status(404).send("Book not found!");
+  }
   const deletedBook = await BookModel.deleteOne({ id });
 
-  return deletedBook;
+  return res
+    .status(201)
+    .send({ msg: "Successfully deleted", data: deletedBook });
 }
 
+async function addBookAuthor(req, res) {
+  const { id } = req.params;
+  const author = getAuthorById(id);
 
-module.exports = { getBookById, updateBook, deleteBook };
+  if (!author) {
+    return res.status(404).send("Author not found!");
+  }
+
+  const { book } = req.body;
+  await author.books.save(book);
+
+  return res.status(201).send("Added book");
+}
+module.exports = { getBooks, updateBook, deleteBook, addBookAuthor };
